refactor: move SettingsSwitch to common
