+++ packages/webapp/pages/account/appearance.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback } from 'react';
-import type { ReactElement, ReactNode } from 'react';
+import type { ReactElement } from 'react';
 import type { NextSeoProps } from 'next-seo';
 
 import { ThemeSection } from '@dailydotdev/shared/src/components/ProfileMenu/sections/ThemeSection';
@@ -10,7 +10,6 @@ import {
   TypographyColor,
   TypographyType,
 } from '@dailydotdev/shared/src/components/typography/Typography';
-import { Switch } from '@dailydotdev/shared/src/components/fields/Switch';
 import { Radio } from '@dailydotdev/shared/src/components/fields/Radio';
 import { ToggleRadio } from '@dailydotdev/shared/src/components/fields/ToggleRadio';
 import { useLogContext } from '@dailydotdev/shared/src/contexts/LogContext';
@@ -24,33 +23,7 @@ import { AccountPageContainer } from '../../components/layouts/AccountLayout/Acc
 import { getAccountLayout } from '../../components/layouts/AccountLayout';
 import { defaultSeo } from '../../next-seo';
 import { getTemplatedTitle } from '../../components/layouts/utils';
-
-type SettingsSwitchProps = {
-  name?: string;
-  children: ReactNode;
-  checked: boolean;
-  onToggle: () => void;
-};
-
-const SettingsSwitch = ({ name, children, ...props }: SettingsSwitchProps) => {
-  return (
-    <div className="flex justify-between gap-4">
-      <Typography
-        type={TypographyType.Callout}
-        color={TypographyColor.Secondary}
-        className="flex-1"
-      >
-        {children}
-      </Typography>
-      <Switch
-        inputId={`${name}-switch`}
-        name={name}
-        compact={false}
-        {...props}
-      />
-    </div>
-  );
-};
+import { SettingsSwitch } from '../../components/layouts/AccountLayout/common';
 
 const densities = [
   { label: 'Eco', value: 'eco' },
